fix shareed clips

+++ events/message.js
@@ -29,6 +29,7 @@ module.exports = async (client, message) => {
     let shareLinks = ["twitch.tv/", "youtube.com/"]
     try {
         if (message.author.bot) return;
+        if (message.channel.id == '749584245457944577' && message.content.includes('clips.twitch')) return;
         if (message.channel.id !== '746777186714779770' && shareLinks.some(word => message.content.toLowerCase().includes(word))) {
             await message.delete();
             return await message.channel.send(new MessageEmbed()
